improve codebase
