+++ src/frontend/src/Redux/action/ApproveRequest.js
@@ -14,9 +14,15 @@ async(data,{rejectWithValue})=>{
         {repo.Err.Error && ToastError(repo.Err.Error)}
         {repo.Err.InvalidPayload && ToastError(repo.Err.InvalidPayload)}
         {repo.Err.NoProfile&& ( ToastError(repo.Err.NoProfile),
-            setTimeout(()=>{window.location.href="/"}, 3000))}
-            {repo.Err.Unauthorized&& ( ToastError(repo.Err.Unauthorized),
-                setTimeout(()=>{window.location.href="/"}, 3000))} 
+            setTimeout(async()=>{
+                await logout()
+              window.location.href="/"}, 
+              3000))} 
+            {repo.Err.Unauthorized&& (ToastError(repo.Err.Unauthorized),
+                setTimeout(async()=>{
+                    await logout()
+                  window.location.href="/"}, 
+                  3000))} 
     
         return rejectWithValue(repo.Err)
        }
