+++ src/frontend/src/Redux/action/GetAllProfile.js
@@ -1,5 +1,5 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
-import { ToastError, ToastSuccess } from "../../utils/toast";
+import { ToastError } from "../../utils/toast";
 import { GetAllProfile } from "../../utils/endpoints";
 
 export const GetAllProfileThunk = createAsyncThunk("GetAllProfile",
@@ -7,7 +7,6 @@ async(data,{rejectWithValue})=>{
     try{
        const repo = await GetAllProfile() ;
        if(repo.Ok){
-        console.log(repo.Ok)
         return repo.Ok
        }else if(repo.Err){
         {repo.Err.Error && ToastError(repo.Err.Error)}
