+++ src/frontend/src/Redux/action/GetAllStock.js
@@ -1,5 +1,5 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
-import { ToastError, ToastSuccess } from "../../utils/toast";
+import { ToastError } from "../../utils/toast";
 import { GetAllStock } from "../../utils/endpoints";
 
 export const GetAllStockThunk = createAsyncThunk("GetAllStock",
@@ -7,7 +7,6 @@ async(data,{rejectWithValue})=>{
     try{
        const repo = await GetAllStock() ;
        if(repo.Ok){
-        console.log(repo.Ok)
         return repo.Ok
        }else if(repo.Err){
         {repo.Err.Error && ToastError(repo.Err.Error)}
