+++ src/frontend/src/Redux/action/ProfileStats.js
@@ -1,16 +1,15 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import { ToastError } from "../../utils/toast";
-import { GetAllProgram } from "@/utils/endpoints";
+import { userProfileStats } from "@/utils/endpoints";
 
-export const GetAllProgramThunk = createAsyncThunk("GetAllPrograms",
+export const ProfileStatsThunk = createAsyncThunk("userProfileStats",
 async(data,{rejectWithValue})=>{
     try{
-       const repo = await GetAllProgram() ;
+       const repo = await userProfileStats();
        if(repo.Ok){
         return repo.Ok
        }else if(repo.Err){
         {repo.Err.Error && ToastError(repo.Err.Error)}
-        {repo.Err.NotFound && ToastError(repo.Err.NotFound)}
         return rejectWithValue(repo.Err)
        }
