+++ src/frontend/src/Redux/action/ProgramCitizens.js
@@ -1,12 +1,12 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
-import { ToastError, ToastSuccess } from "../../utils/toast";
+import { ToastError } from "../../utils/toast";
 import { ProgramCitizens } from "@/utils/endpoints";
 
 export const ProgramCitizensThunk = createAsyncThunk("ProgramCitizens",
 async(data,{rejectWithValue})=>{
     try{
         const {ProgramId} = data
-       const repo = await ProgramCitizens(ProgramId) ;
+       const repo = await ProgramCitizens(ProgramId);
        if(repo.Ok){
         return repo.Ok
        }else if(repo.Err){
