+++ src/frontend/src/Redux/action/Transfer.js
@@ -1,6 +1,7 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import { ToastError, ToastSuccess } from "../../utils/toast";
 import  { Transfer } from "../../utils/endpoints"
+import { logout } from "@/utils/auth";
 
 export const TransferThunk = createAsyncThunk("Transfer",
 async(data,{rejectWithValue})=>{
@@ -12,11 +13,16 @@ async(data,{rejectWithValue})=>{
        }else if(repo.Err){
         {repo.Err.Error && ToastError(repo.Err.Error)}
         {repo.Err.InvalidPayload && ToastError(repo.Err.InvalidPayload)}
-        {repo.Err.NoProfile&& ( ToastError(repo.Err.NoProfile),
-            setTimeout(()=>{window.location.href="/"}, 3000))}
-        {repo.Err.Unauthorized&& ( ToastError(repo.Err.Unauthorized),
-            setTimeout(()=>{window.location.href="/"}, 3000))} 
-
+        {repo.Err.NoProfile&& (ToastError(repo.Err.NoProfile),
+            setTimeout(async()=>{
+                await logout()
+                window.location.href="/"
+            }, 3000))}
+        {repo.Err.Unauthorized&& (ToastError(repo.Err.Unauthorized),
+            setTimeout(async()=>{
+                  await logout()
+                window.location.href="/"}, 
+                3000))} 
         return rejectWithValue(repo.Err)
        }
